Change enclosing scope to name scope
Also, for the function get_output_for in ParamLayer and OpLayer, a scope
to encapsulate the operations performed in the layer has been inserted.
This also needs to be done for other layers in garage/tf/core/layers.py.
For the function get_output in the layers.py script, a name_scope is
required before calling the function to encapsulate all the layer
operation scopes (see garage/tf/policies/gaussian_mlp_policy.py).